requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-admin-honeypot==1.1.0
django-admin-interface==0.19.1
django-admin-sortable2==2.0.4
django-colorfield==0.7.0
django-cors-headers==3.12.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.12
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
